+++ src/components/services/calculateFootprintPerTransportationType.js
