+++ src/components/services/calculateFootprintPerTransportationType.test.js
