+++ src/components/services/calculateTotalFootprintSum.js
