+++ src/components/services/calculateTotalFootprintSum.test.js
