+++ src/components/services/mappings/mapApiTransportationTypeToDonutChartLabel.js
