+++ src/components/services/mappings/mapApiTransportationTypeToDonutChartLabel.test.js
