+++ src/components/services/mappings/mapFootprintPerTransportTypeToDonutChartData.js
