+++ src/components/services/mappings/mapFootprintPerTransportTypeToDonutChartData.test.js
