+++ src/components/services/useDeviceDetect.js
