+++ src/pages/SportsActivitiesPage/SportsActivitiesPage.js
@@ -3,7 +3,7 @@ import { toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import styled from 'styled-components'
 import AddKilometersForm from '../../components/AddKilometers/AddKilometersForm'
-import { calculateCarbonEmission } from '../../components/utils/calculateCarbonEmission'
+import { calculateCarbonEmission } from '../../components/services/calculateCarbonEmission'
 
 export default function SportsActivitiesPage() {
   const [sportsType, setSportsType] = useState('swim')
